@@ -26,35 +26,36 @@
         "v8Flags": "--expose_gc"
     },
     "dependencies": {
-        "@angular/animations": "~5.2.0",
-        "@angular/common": "~5.2.0",
-        "@angular/compiler": "~5.2.0",
-        "@angular/core": "~5.2.0",
-        "@angular/forms": "~5.2.0",
-        "@angular/http": "~5.2.0",
-        "@angular/platform-browser": "~5.2.0",
-        "@angular/platform-browser-dynamic": "~5.2.0",
-        "@angular/router": "~5.2.0",
-        "kinvey-nativescript-sdk": "3.9.7",
-        "nativescript-angular": "~5.3.0",
+        "@angular/animations": "~6.0.0",
+        "@angular/common": "~6.0.0",
+        "@angular/compiler": "~6.0.0",
+        "@angular/core": "~6.0.0",
+        "@angular/forms": "~6.0.0",
+        "@angular/http": "~6.0.0",
+        "@angular/platform-browser": "~6.0.0",
+        "@angular/platform-browser-dynamic": "~6.0.0",
+        "@angular/router": "~6.0.0",
+        "kinvey-nativescript-sdk": "3.10.3",
+        "nativescript-angular": "next",
         "nativescript-email": "~1.5.1",
-        "nativescript-iqkeyboardmanager": "~1.1.0",
+        "nativescript-iqkeyboardmanager": "~1.3.0",
         "nativescript-phone": "~1.3.1",
-        "nativescript-push-notifications": "~1.0.0",
         "nativescript-theme-core": "~1.0.4",
-        "nativescript-ui-dataform": "~3.5.0",
-        "nativescript-ui-gauge": "~3.5.0",
+        "nativescript-ui-dataform": "~3.6.0",
+        "nativescript-ui-gauge": "~3.6.0",
         "nativescript-ui-listview": "~3.5.0",
         "reflect-metadata": "~0.1.10",
-        "rxjs": "~5.5.5",
+        "rxjs": "~6.1.0",
+        "rxjs-compat": "~6.1.0",
+        "tns-core-modules": "next",
         "zone.js": "~0.8.18"
     },
     "devDependencies": {
-        "codelyzer": "~4.0.2",
+        "codelyzer": "~4.3.0",
         "nativescript-dev-typescript": "~0.7.0",
-        "nativescript-dev-webpack": "~0.10.0",
+        "nativescript-dev-webpack": "next",
         "tns-platform-declarations": "~4.0.0",
-        "tslint": "~5.8.0",
+        "tslint": "~5.10.0",
         "typescript": "~2.7.2"
     },
     "scripts": {
